Add query_converter module
